pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "clockwork"
version = "0.2.0"
description = "Local-first CI/CD with self-healing pipelines"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
license-files = ["LICENSE"]

dependencies = [
    "typer>=0.12",
    "pyyaml>=6.0",
    "gitpython>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov",
]
graph = [
    "networkx>=3.0",
]
mcp = [
    "mcp>=1.0",
]
http = [
    "aiohttp>=3.9",
]
all = [
    "mcp>=1.0",
    "aiohttp>=3.9",
    "watchdog>=4.0",
    "questionary>=2.0",
]

[project.scripts]
clockwork = "clockwork.cli.app:main"
clockwork-mcp = "clockwork.mcp_server:main"

[project.urls]
Homepage = "https://github.com/ShibilAhamed701212/clockwork"
Repository = "https://github.com/ShibilAhamed701212/clockwork"

[tool.setuptools.packages.find]
where = ["."]
include = ["clockwork*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
pythonpath = ["."]
addopts = "-ra"

[tool.coverage.run]
source = ["clockwork"]
omit = ["tests/*"]
